+++ CHANGELOG.md
@@ -1,11 +1,26 @@
-#### 0.1.7 (2021-12-10)
+#### 0.2.0-recordings.1 (2021-12-14)
+
+##### Continuous Integration
+
+* **release:**  bump up version: 0.2.0-recordings.0 (12d19688)
 
 ##### New Features
 
-*  bg pic for login (714238ab)
+*  gzip batch data (d10005a5)
+*  save openFDR data to disk (d6cf44e8)
+
+##### Bug Fixes
+
+*  deduplicate ws ping (c3d74d27)
+*  avoid dup final batch (f2f97416)
 
 ##### Other Changes
 
-* //github.com/xairline/zon (2c9a44e4)
+* //github.com/xairline/zon into openfdr (90dc8aa1)
+* //github.com/xairline/zon into openfdr (bdd25088)
+* //github.com/xairline/zon into openfdr (3c09e042)
+* //github.com/xairline/zon into openfdr (aeedd4d5)
+* //github.com/xairline/zon into openfdr (c3fce1e2)
+* //github.com/xairline/zon into openfdr (32ddc7fa)